Move config back into filecheck temporarily
